p105: The Skyline Problem
Oh what a wonderful feeling... I had attempted to solve this problem
for the first time 2007-12-31 15:47:30 and after 12 unsuccessful
submissions I gave up at 2008-01-01 09:08:01.

10 years and 6 months later I gave it another shot and finally got it
right.

That means that I am not getting stupider as I get older, or at least
I've learned a few tricks along the way that make up for it.