p103: Stacking Boxes
Another problem that I tried for the first time 2007-11-02 18:27:24
and then gave up at 2007-11-02 22:27:52, barely 6 hours later.

Now I've solved it in a little bit under 3 hours, which is not a
great time for competitive standards, but it's still nice to see
that I am performing better than 10 years ago.